Huge sweeping changes.
First time through, I didn't have enough in the way of testing so this
is an initial stab at fixing that. Several unneccessary features are
also being phased out. Including framing and matting even though a lot
of the initial work here was in fixing those to work more logically.