Don't need to manage users as the admin.
Futhermore, due to the nature of the in memory test database,
this was causing User to be loaded before the table existed in our
database. This caused authlogic to not set up currectly since it
reflects on columns in the databse.

Easy fix at this point is to just not use active admin for users instead
of fixing the order of files loading and when database is configured.